Allow `summarize` and `effect-size` to produce human-readable output
Previously, only the `benchmark` command understood how to print the
summarized results to a human-readable output. The `summarize` and
`effect-size` commands only could print to CSV or JSON. While
benchmarking a scenario that required separate runs and thus separate
result files, I realized that a) I had to reduce the various result
files to a single file and b) there was no way to print a human-readable
version of the summaries or effect sizes. This change refactors the
implementation of the human-readable display to the sightglass-analysis
crate so it can be used in multiple commands (`benchmark`, `summarize`,
`effect-size`). The human-readable output is used if the user does not
specify a (now optional) `--output-format`. This change also makes it
possible to pass multiple files to `summarize` and `effect-size`
commands
