+++ crates/analysis/src/effect_size.rs
@@ -1,7 +1,7 @@
-use crate::KeyBuilder;
+use crate::keys::KeyBuilder;
 use anyhow::Result;
-use sightglass_data::{EffectSize, Measurement};
-use std::collections::BTreeSet;
+use sightglass_data::{EffectSize, Measurement, Phase, Summary};
+use std::{collections::BTreeSet, io::Write};
 
 /// Find the effect size (and confidence interval) of between two different
 /// engines (i.e. two different commits of Wasmtime).
@@ -13,7 +13,7 @@ use std::collections::BTreeSet;
 /// This can only test differences between the results for exactly two different
 /// engines. If there aren't exactly two different engines represented in
 /// `measurements` then an error is returned.
-pub fn effect_size<'a>(
+pub fn calculate<'a>(
     significance_level: f64,
     measurements: &[Measurement<'a>],
 ) -> Result<Vec<EffectSize<'a>>> {
@@ -71,4 +71,123 @@ pub fn effect_size<'a>(
     }
 
     Ok(results)
+}
+
+/// Write a vector of [EffectSize] structures to the passed `output_file` in human-readable form.
+/// The `summaries` are needed
+pub fn write(
+    mut effect_sizes: Vec<EffectSize<'_>>,
+    summaries: &[Summary<'_>],
+    significance_level: f64,
+    output_file: &mut dyn Write,
+) -> Result<()> {
+    // Sort the effect sizes so that we focus on statistically significant results before
+    // insignificant results and larger relative effect sizes before smaller relative effect sizes.
+    effect_sizes.sort_by(|x, y| {
+        y.is_significant().cmp(&x.is_significant()).then_with(|| {
+            let x_speedup = x.a_speed_up_over_b().0.max(x.b_speed_up_over_a().0);
+            let y_speedup = y.a_speed_up_over_b().0.max(y.b_speed_up_over_a().0);
+            y_speedup.partial_cmp(&x_speedup).unwrap()
+        })
+    });
+
+    for effect_size in effect_sizes {
+        writeln!(output_file)?;
+        writeln!(
+            output_file,
+            "{} :: {} :: {}",
+            effect_size.phase, effect_size.event, effect_size.wasm
+        )?;
+        writeln!(output_file)?;
+
+        // For readability, trim the shared prefix from our two engine names.
+        let end_of_shared_prefix = effect_size
+            .a_engine
+            .char_indices()
+            .zip(effect_size.b_engine.char_indices())
+            .find_map(|((i, a), (j, b))| {
+                if a == b {
+                    None
+                } else {
+                    debug_assert_eq!(i, j);
+                    Some(i)
+                }
+            })
+            .unwrap_or(0);
+        let a_engine = &effect_size.a_engine[end_of_shared_prefix..];
+        let b_engine = &effect_size.b_engine[end_of_shared_prefix..];
+
+        if effect_size.is_significant() {
+            writeln!(
+                output_file,
+                "  Δ = {:.2} ± {:.2} (confidence = {}%)",
+                (effect_size.b_mean - effect_size.a_mean).abs(),
+                effect_size.half_width_confidence_interval.abs(),
+                (1.0 - significance_level) * 100.0,
+            )?;
+            writeln!(output_file)?;
+
+            let ratio = effect_size.b_mean / effect_size.a_mean;
+            let ratio_ci = effect_size.half_width_confidence_interval / effect_size.a_mean;
+            writeln!(
+                output_file,
+                "  {a_engine} is {ratio_min:.2}x to {ratio_max:.2}x faster than {b_engine}!",
+                a_engine = a_engine,
+                b_engine = b_engine,
+                ratio_min = ratio - ratio_ci,
+                ratio_max = ratio + ratio_ci,
+            )?;
+            let ratio = effect_size.a_mean / effect_size.b_mean;
+            let ratio_ci = effect_size.half_width_confidence_interval / effect_size.b_mean;
+
+            writeln!(
+                output_file,
+                "  {b_engine} is {ratio_min:.2}x to {ratio_max:.2}x faster than {a_engine}!",
+                a_engine = a_engine,
+                b_engine = b_engine,
+                ratio_min = ratio - ratio_ci,
+                ratio_max = ratio + ratio_ci,
+            )?;
+        } else {
+            writeln!(output_file, "  No difference in performance.")?;
+        }
+        writeln!(output_file)?;
+
+        let get_summary = |engine: &str, wasm: &str, phase: Phase, event: &str| {
+            // TODO this sorting is not using `arch` which is not guaranteed to be the same in
+            // result sets; potentially this could re-use `Key` functionality.
+            summaries
+                .iter()
+                .find(|s| {
+                    s.engine == engine && s.wasm == wasm && s.phase == phase && s.event == event
+                })
+                .unwrap()
+        };
+
+        let a_summary = get_summary(
+            &effect_size.a_engine,
+            &effect_size.wasm,
+            effect_size.phase,
+            &effect_size.event,
+        );
+        writeln!(
+            output_file,
+            "  [{} {:.2} {}] {}",
+            a_summary.min, a_summary.mean, a_summary.max, a_engine,
+        )?;
+
+        let b_summary = get_summary(
+            &effect_size.b_engine,
+            &effect_size.wasm,
+            effect_size.phase,
+            &effect_size.event,
+        );
+        writeln!(
+            output_file,
+            "  [{} {:.2} {}] {}",
+            b_summary.min, b_summary.mean, b_summary.max, b_engine,
+        )?;
+    }
+
+    Ok(())
 }
