+++ crates/analysis/src/lib.rs
@@ -1,9 +1,3 @@
-mod effect_size;
-mod keys;
-mod summarize;
-
-pub use {
-    effect_size::effect_size,
-    keys::{Key, KeyBuilder},
-    summarize::summarize,
-};
+pub mod effect_size;
+pub mod keys;
+pub mod summarize;
